requirements.txt
nvidia-riva-client>=2.14.0
openai>=1.0.0
python-dotenv>=1.0.0
soundfile>=0.12.0
numpy>=1.21.0
scipy>=1.7.0
pyaudio>=0.2.14
jupyter>=1.0.0
jupyterlab>=3.0.0
pytest>=7.0.0
grpcio>=1.50.0
protobuf>=3.20.0
